chore: moved module loader to its specific crate
