+++ crates/module_loader/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "module_loader"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+common.path="../common"
+slynx-parser="../parser"
+slynx-lexer="../lexer"
