+++ crates/module_loader/src/lib.rs
@@ -1,3 +1,4 @@
+mod file;
 use std::{
     collections::{HashMap, VecDeque},
     path::{Path, PathBuf},
@@ -7,20 +8,7 @@ use common::Span;
 use slynx_lexer::{Lexer, LexerError};
 use slynx_parser::{ASTDeclaration, ASTDeclarationKind, ASTPath, ParseError, Parser};
 
-#[derive(Debug, Clone, Copy, Hash, PartialEq, PartialOrd, Ord, Eq)]
-///An ID to represent a file
-pub struct FileId(u32);
-impl FileId {
-    pub fn from_raw(value: u32) -> Self {
-        Self(value)
-    }
-    pub fn as_raw(&self) -> u32 {
-        self.0
-    }
-    fn next(&self) -> Self {
-        Self(self.0 + 1)
-    }
-}
+pub struct Modules {}
 
 #[derive(Debug)]
 ///A module that was parsed
